Rearrange folders and rename blank page
