+++ frontend/src/api/adapters/tiingo/TiingoApi.ts
@@ -1,6 +1,21 @@
-import {TiingoApiPriceResponse} from "./tiingo-interfaces";
-import {PriceHistoryItem} from "../../interfaces/goplan-interfaces";
-import {PriceAdapterInterface} from "../../interfaces/adapter-interfaces";
+import {PriceHistoryItem} from "../../../interfaces/ApplicationInterfaces";
+import {PriceAdapterInterface} from "../../../interfaces/AdapterInterfaces";
+
+interface TiingoApiPriceResponse {
+    date: Date;
+    close: number;
+    high: number;
+    low: number;
+    open: number;
+    volume: number;
+    adjClose: number;
+    adjHigh: number;
+    adjLow: number;
+    adjOpen: number;
+    adjVolume: number;
+    divCash: number;
+    splitFactor: number;
+}
 
 export class TiingoApiAdapter implements PriceAdapterInterface {
     async getPrices(ticker: string) {
