+++ frontend/src/components/Sidebar.vue
@@ -179,8 +179,8 @@
 
         <router-link
           class="flex items-center duration-200 mt-4 py-2 px-6 border-l-4"
-          :class="[$route.name === 'Blank' ? activeClass : inactiveClass]"
-          to="/blank"
+          :class="[$route.name === 'Test' ? activeClass : inactiveClass]"
+          to="/test"
         >
           <svg class="h-5 w-5" fill="currentColor" viewBox="0 0 20 20">
             <path
@@ -188,7 +188,7 @@
             />
           </svg>
 
-          <span class="mx-4">Blank</span>
+          <span class="mx-4">Test</span>
         </router-link>
       </nav>
     </div>
