--- frontend/src/hooks/api/adapters/tiingo/tiingo-interfaces.ts
@@ -1,15 +1,0 @@
-export interface TiingoApiPriceResponse {
-    date: Date;
-    close: number;
-    high: number;
-    low: number;
-    open: number;
-    volume: number;
-    adjClose: number;
-    adjHigh: number;
-    adjLow: number;
-    adjOpen: number;
-    adjVolume: number;
-    divCash: number;
-    splitFactor: number;
-}
