+++ frontend/src/interfaces/ApplicationInterfaces.ts
